Split STV tests intp multiple files to make editors less laggy
